reorganize header components
