+++ components/header/audio-waveform/audio-waveform-live.tsx
