+++ components/header/audio-waveform/audio-waveform-offline.tsx
