+++ components/header/audio-waveform/index.tsx
